% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accio.R
\name{accio}
\alias{accio}
\title{Install, update and attach packages}
\usage{
accio(packages, update = FALSE, silent = TRUE)
}
\arguments{
\item{packages}{A character vector. The names of the packages to attach or
install.}

\item{update}{A logical scalar. Should outdated packages be updated as well?}

\item{silent}{A logical scalar. Suppress all errors, warnings, and messages?}
}
\value{
A logical vector of the same lenght as \code{packages}. Each entry
indicates whether loading the package was successful.
}
\description{
A utility function for easy installation and attaching of packages. If a
package is unavailable in any of the local libraries, it will be fetched and
installed.
}
\examples{
accio(c("tictoc", "praise", "available", "xxyyzz"))
}
