% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ayuda.R
\name{ayuda}
\alias{ayuda}
\title{Quickly View of PDF Help}
\source{
\href{https://stackoverflow.com/questions/30607496/making-an-r-package-pdf-manual-using-devtools}{This}
stackoverflow answer.
}
\usage{
ayuda(topic)
}
\arguments{
\item{topic}{A name, the topic for which to produce a PDF. Must not be a
character string (see examples).}
}
\value{
NULL. Called for its side-effect, which is to create a PDF at a
temporary location and open it.
}
\description{
Una \ifelse{latex}{\out{funci{\'o}n}}{función} por su ayuda. View the PDF
help of a package or function.
}
\note{
Requires the texi2dvi utility. It can be found in the \emph{texinfo} package
on debian and derivatives.
}
\examples{

\dontrun{

ayuda(utils::help)
ayuda(utils)

ayuda(charms::ayuda)
}

}
