% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressio.R
\name{prgr}
\alias{prgr}
\title{Title}
\usage{
prgr(fun, iterations)
}
\arguments{
\item{iterations}{}
}
\description{
Title
}
\examples{

shout <- function() Sys.sleep(0.1)

reps <- 100
shout <- prgr(shout, reps)

Sys.sleep(3)
replicate(reps, shout())

cl <- parallel::makeCluster(2)
fun <- function(x) {Sys.sleep(1); x}
system.time(
  parallel::parLapply(cl, 1:10, fun)
)

fun <- prgr(fun, iterations = 10)
system.time(
  parallel::clusterCall(cl, fun)
)

doParallel::registerDoParallel(cl)

foreach::`\%do\%`(foreach::foreach(1:10, .packages = "tcltk"), fun())

foreach::foreach()

}
