% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressio.R
\name{progressio}
\alias{progressio}
\title{Title}
\usage{
progressio(iterations, title = "Progress")
}
\arguments{
\item{iterations}{}
}
\description{
Title
}
\examples{


progrede <- progressio(100)
replicate(100, progrede())


progrede <- progressio(100)
replicate(100, {Sys.sleep(0.05); progrede()})


TODO: option to output progress on command line
}
