% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silencio.R
\name{silencio}
\alias{silencio}
\alias{sonorus}
\title{Silence All Output}
\usage{
silencio()

sonorus()
}
\description{
Silence all output including messages, warnings and errors that would
otherwise be printed to the console or terminal. Use your \code{sonorus()} charm
to disable.
}
\details{
If you dislike Harry Potter, you would have to play a round of \code{hide()} and
\code{seek()} instead. However, these are not exported to avoid confusion with
\code{\link[base:seek]{base::seek()}}.
}
\examples{
warning("You!")

#---------
silencio()
#---------

warning("You!")
stop("You, again!")

#--------
sonorus()
#--------

message("finally, it stopped")

charms:::hide()
print("this works, too?")
charms:::seek()
cat("yes")
}
